% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_template.R
\name{raster_template}
\alias{raster_template}
\title{Create raster template}
\usage{
raster_template(
  ext,
  cellsize,
  offset = NULL,
  crs = "+proj=longlat +datum=WGS84"
)
}
\arguments{
\item{ext}{An object of class extent or a vector contain xmin xmax ymin ymax}

\item{cellsize}{Output cell of raster, either a one or two element vector}

\item{offset}{A one or two element vector providing values for how raster
should be shifted in the x and y direction}

\item{crs}{CRS of output raster, defaulting to EPSG:4326}
}
\value{
A raster with values from 1:ncell
}
\description{
Create raster template
}
\details{
A convenience function to allow easy creation of rasters with
arbitrary cell size, for populating with other values.
with other
}
\examples{
e <- c(25, 27, -14, -12)
r <- raster_template(e, cellsize = 0.05)
r2 <- raster_template(raster::extent(r), cellsize = 0.1)
r3 <- raster_template(raster::extent(r), offset = 0.05, cellsize = 0.1)
par(mfrow = c(1, 3))
for(i in list(r, r2, r3)) raster::plot(i)
}
